iniciando testes

+++ LocadoraV1semDB/LocadoraBackEnd/src/main/java/com/grupo12/Entidades/Descontos/RegraDesconto.java
@@ -1,5 +1,5 @@
 package com.grupo12.Entidades.Descontos;
 
 public interface RegraDesconto {
-    public int disconto();
+    public double disconto();
 }
